OS-4 feat(cart): add item description link to items
